Remove duplicate codes
